+++ cratedb_toolkit/io/omniload/api.py
@@ -3,51 +3,51 @@ from functools import lru_cache
 
 from yarl import URL
 
-from cratedb_toolkit.io.ingestr.boot import import_ingestr
+from cratedb_toolkit.io.omniload.boot import import_omniload
 from cratedb_toolkit.model import DatabaseAddress
 
 logger = logging.getLogger(__name__)
 
 
 @lru_cache(maxsize=1)
-def _get_ingestr():
+def _get_omniload():
     """
-    Get an ingestr API handle, cached.
+    Get an omniload API handle, cached.
     """
-    return import_ingestr()
+    return import_omniload()
 
 
-def ingestr_select(source_url: str) -> bool:
+def omniload_select(source_url: str) -> bool:
     """
-    Whether to select `ingestr` for this data source.
+    Whether to select `omniload` for this data source.
     """
-    ingestr_available, ingestr, ConfigFieldMissingException = _get_ingestr()
+    omniload_available, omniload, ConfigFieldMissingException = _get_omniload()
 
-    if not ingestr_available:
-        logger.debug("ingestr is not installed")
+    if not omniload_available:
+        logger.debug("omniload is not installed")
         return False
     try:
-        factory = ingestr.src.factory.SourceDestinationFactory(source_url, "csv:////tmp/foobar.csv")
+        factory = omniload.src.factory.SourceDestinationFactory(source_url, "csv:////tmp/foobar.csv")
         factory.get_source()
-        scheme = ingestr.src.factory.parse_scheme_from_uri(source_url)
-        logger.info(f"Selecting ingestr for source scheme: {scheme}")
+        scheme = omniload.src.factory.parse_scheme_from_uri(source_url)
+        logger.info(f"Selecting omniload for source scheme: {scheme}")
         return True
     except (ImportError, ValueError, AttributeError) as ex:
         if "Unsupported source scheme" in str(ex):
-            logger.debug(f"Failed to select ingestr for source url '{source_url}': {ex}")
+            logger.debug(f"Failed to select omniload for source url '{source_url}': {ex}")
         else:
-            logger.exception(f"Unexpected error with ingestr for source url: {source_url}")
+            logger.exception(f"Unexpected error with omniload for source url: {source_url}")
         return False
     except Exception:
-        logger.exception(f"Unexpected error with ingestr for source url: {source_url}")
+        logger.exception(f"Unexpected error with omniload for source url: {source_url}")
         return False
 
 
-def ingestr_copy(source_url: str, target_address: DatabaseAddress, progress: bool = False) -> bool:
+def omniload_copy(source_url: str, target_address: DatabaseAddress, progress: bool = False) -> bool:
     """
-    Invoke data transfer to CrateDB from any source provided by `ingestr`.
+    Invoke data transfer to CrateDB from any source provided by `omniload`.
 
-    https://cratedb-toolkit.readthedocs.io/io/ingestr/
+    https://cratedb-toolkit.readthedocs.io/io/omniload/
 
     Synopsis:
 
@@ -63,11 +63,11 @@ def ingestr_copy(source_url: str, target_address: DatabaseAddress, progress: boo
             "postgresql://pguser:secret11@postgresql.example.org:5432/postgres?table=public.diamonds" \
             "crate://crate:na@localhost:4200/testdrive/ibis_diamonds"
     """
-    ingestr_available, ingestr, ConfigFieldMissingException = _get_ingestr()
+    omniload_available, omniload, ConfigFieldMissingException = _get_omniload()
 
     # Sanity checks.
-    if not ingestr_available:
-        raise ModuleNotFoundError("ingestr subsystem not installed")
+    if not omniload_available:
+        raise ModuleNotFoundError("omniload subsystem not installed")
 
     # Compute source and target URLs and table names.
     # Table names use dotted notation `<schema>.<table>`.
@@ -91,7 +91,7 @@ def ingestr_copy(source_url: str, target_address: DatabaseAddress, progress: boo
 
     target_uri, target_table_address = target_address.decode()
     target_table = target_table_address.fullname
-    target_url = target_address.to_ingestr_url()
+    target_url = target_address.to_omniload_url()
 
     if not source_table:
         raise ValueError("Source table is required")
@@ -101,7 +101,7 @@ def ingestr_copy(source_url: str, target_address: DatabaseAddress, progress: boo
     if source_fragment:
         source_table += f"#{source_fragment}"
 
-    logger.info("Invoking ingestr")
+    logger.info("Invoking omniload")
     logger.info(f"Source URL: {source_url_obj}")
     logger.info(f"Target URL: {target_url}")
     logger.info(f"Source Table: {source_table}")
@@ -114,7 +114,6 @@ def ingestr_copy(source_url: str, target_address: DatabaseAddress, progress: boo
         dest_uri=str(target_url),
         source_table=source_table,
         dest_table=target_table,
-        yes=True,
     )
     if start_date is not None:
         ingest_kwargs["interval_start"] = start_date
@@ -122,7 +121,7 @@ def ingestr_copy(source_url: str, target_address: DatabaseAddress, progress: boo
         ingest_kwargs["page_size"] = batch_size
 
     try:
-        ingestr.main.ingest(**ingest_kwargs)
+        omniload.main.ingest(**ingest_kwargs)
         return True
     except ConfigFieldMissingException:
         logger.error(