+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Fixed a failing SQL statement during `ctk info cluster`
 - Dependency: Update sqlalchemy-crate version to 0.43.1
 - Removed support for Rockset
+- I/O: Migrated from `ingestr` to `omniload`
 
 ## 2026/06/17 v0.0.49
 - Stopped leaking password to log output in `ctk cfr jobstats collect`.